+++ apps/wizard/app_pages/indicator_upgrade/indicator_mapping.py
@@ -7,7 +7,6 @@ import pandas as pd
 import streamlit as st
 from structlog import get_logger
 
-from apps.backport.datasync.data_metadata import variable_data_df_from_s3
 from apps.wizard.app_pages.indicator_upgrade.explore_mode import st_explore_indicator
 from apps.wizard.app_pages.indicator_upgrade.utils import (
     find_mapping_suggestions_cached,
@@ -17,6 +16,7 @@ from apps.wizard.app_pages.indicator_upgrade.utils import (
 from apps.wizard.utils import Pagination, set_states
 from etl.config import OWID_ENV
 from etl.db import get_engine, read_sql
+from etl.grapher_io import variable_data_df_from_s3
 
 # Logger
 log = get_logger()
