+++ etl/compare.py
@@ -16,9 +16,9 @@ from rich import print
 from rich_click.rich_command import RichCommand
 from rich_click.rich_group import RichGroup
 
-from apps.backport.datasync.data_metadata import variable_data_df_from_s3
 from etl import tempcompare
 from etl.db import get_engine, read_sql
+from etl.grapher_io import variable_data_df_from_s3
 
 
 @click.group(name="compare", cls=RichGroup)
