+++ etl/scripts/faostat/archive/migrate_to_new_metadata.py
@@ -11,8 +11,8 @@ import json
 from owid import catalog
 from structlog import get_logger
 
-from etl import db
 from etl.files import yaml_dump
+from etl.grapher_io import get_dataset_id, get_variables_in_dataset
 from etl.paths import DATA_DIR, STEP_DIR
 
 # Initialize logger.
@@ -44,14 +44,14 @@ def main():
             )
             continue
         try:
-            dataset_id = db.get_dataset_id(dataset_name=dataset_name, version=VERSION)  # type: ignore[reportArgumentType]
+            dataset_id = get_dataset_id(dataset_name=dataset_name, version=VERSION)  # type: ignore[reportArgumentType]
         except AssertionError:
             log.error(
                 f"Grapher dataset for {domain} could not be found in the database. "
                 f"Run `etl run {domain} --grapher` and try again."
             )
             continue
-        variables = db.get_variables_in_dataset(dataset_id=dataset_id, only_used_in_charts=True)
+        variables = get_variables_in_dataset(dataset_id=dataset_id, only_used_in_charts=True)
         if len(variables) > 0:
             # variables_in_charts[domain] = variables.set_index("shortName").T.to_dict()
             variables_dict = {}
